+++ test/test-write.c
@@ -49,7 +49,7 @@ TEST (write, msr3_writemseed_encodings)
   int32_t isinedata[SINE_DATA_SAMPLES];
   float fsinedata[SINE_DATA_SAMPLES];
   int idx;
-  int rv;
+  int64_t rv;
 
   /* Create integer and double sine data sets */
   for (idx = 0; idx < SINE_DATA_SAMPLES; idx++)
@@ -257,7 +257,7 @@ TEST (write, msr3_writemseed_headeronly_v2)
 {
   MS3Record *msr = NULL;
   uint32_t flags = MSF_PACKVER2; /* write v2 format */
-  int rv;
+  int64_t rv;
 
   msr = msr3_init (msr);
   REQUIRE (msr != NULL, "msr3_init() returned unexpected NULL");
@@ -272,7 +272,7 @@ TEST (write, msr3_writemseed_headeronly_v2)
   msr->pubversion = 1;
 
   msr->extra = soh_json_headers;
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   msr->samplecnt = 0;
   msr->numsamples = 0;
@@ -293,7 +293,7 @@ TEST (write, msr3_writemseed_headeronly_v3)
 {
   MS3Record *msr = NULL;
   uint32_t flags = 0;
-  int rv;
+  int64_t rv;
 
   msr = msr3_init (msr);
   REQUIRE (msr != NULL, "msr3_init() returned unexpected NULL");
@@ -308,7 +308,7 @@ TEST (write, msr3_writemseed_headeronly_v3)
   msr->pubversion = 1;
 
   msr->extra = soh_json_headers;
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   msr->samplecnt = 0;
   msr->numsamples = 0;
@@ -334,7 +334,7 @@ TEST (write, msr3_writemseed_nanosecond)
   uint32_t flags = MSF_FLUSHDATA; /* Set data flush flag */
   int32_t isinedata[SINE_DATA_SAMPLES];
   int idx;
-  int rv;
+  int64_t rv;
 
   /* Create integer sine data set */
   for (idx = 0; idx < SINE_DATA_SAMPLES; idx++)
@@ -370,7 +370,7 @@ TEST (write, msr3_writemseed_nanosecond)
                      "\"Clock\":{"
                      "\"Model\":\"Acme Corporation GPS3\""
                      "}}}";
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   rv = msr3_writemseed (msr, TESTFILE_NSEC_V3, 1, flags, 0);
   REQUIRE (rv > 0, "msr3_writemseed() return unexpected value");
@@ -397,7 +397,7 @@ TEST (write, msr3_writemseed_nanosecond)
                      "\"Clock\":{"
                      "\"Model\":\"Acme Corporation GPS3\""
                      "}}}";
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   rv = msr3_writemseed (msr, TESTFILE_NSEC_V2, 1, flags, 0);
   REQUIRE (rv > 0, "msr3_writemseed() return unexpected value");
@@ -419,7 +419,7 @@ TEST (write, msr3_writemseed_olden)
   uint32_t flags = MSF_FLUSHDATA; /* Set data flush flag */
   int32_t isinedata[SINE_DATA_SAMPLES];
   int idx;
-  int rv;
+  int64_t rv;
 
   /* Create integer sine data set */
   for (idx = 0; idx < SINE_DATA_SAMPLES; idx++)
@@ -453,7 +453,7 @@ TEST (write, msr3_writemseed_olden)
                      "\"Clock\":{"
                      "\"Model\":\"Ye Olde Clock Tower Company\""
                      "}}}";
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   rv = msr3_writemseed (msr, TESTFILE_OLDEN_V3, 1, flags, 0);
   REQUIRE (rv > 0, "msr3_writemseed() return unexpected value");
@@ -478,7 +478,7 @@ TEST (write, msr3_writemseed_olden)
                      "\"Clock\":{"
                      "\"Model\":\"Ye Olde Clock Tower Company\""
                      "}}}";
-  msr->extralength = strlen (msr->extra);
+  msr->extralength = (uint16_t)strlen (msr->extra);
 
   rv = msr3_writemseed (msr, TESTFILE_OLDEN_V2, 1, flags, 0);
   REQUIRE (rv > 0, "msr3_writemseed() return unexpected value");
@@ -503,7 +503,7 @@ TEST (write, msr3_writemseed_oddrate)
   uint32_t flags = MSF_FLUSHDATA; /* Set data flush flag */
   int32_t isinedata[SINE_DATA_SAMPLES];
   int idx;
-  int rv;
+  int64_t rv;
 
   /* Create integer sine data set */
   for (idx = 0; idx < SINE_DATA_SAMPLES; idx++)
